requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.1
Brotli==1.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
mutagen==1.47.0
propcache==0.3.2
py-cord==2.6.1
pycryptodomex==3.23.0
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
validators==0.35.0
websockets==15.0.1
yarl==1.20.1
